+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.6.0] - 2022-01-30
+
+### Changed
+- `LoudnessToAmplitudeConverter` to `LoudnessToAmplitude`
+- `RhythmicalStrataToIndispensabilityConverter` to `RhythmicalStrataToIndispensability`
+- `TwoPitchesToCommonHarmonicsConverter` to `TwoPitchesToCommonHarmonics`
+
 ## [0.5.0] - 2022-01-30
 
 ### Changed